Allow JSON data from metadata
Instead of supplying a source_path in the plugin options and source_file
names in the json_files object, allow the plugin to use values present
in the Metalsmith metadata as the JSON source.

Two new settings:

1) `use_metadata`: a boolean that if true, will allow the use of
metadata as the json source.
2) `metadata_key: a string in the `json_files` object that is the key
in the Metalsmith metadata where the json source is located.

Only `use_metadata` or `source_path` may be passed as options, not both.

Only `metadata_key` or `source_file` may be present in `json_files` not
both.